pyproject.toml
[project]
name = "substack-archives-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "selenium>=4.0.0",
    "soupsieve>=2.2.1",
    "urllib3>=1.26.18",
    "validators>=0.18.2",
    "webdriver-manager>=4.0.2",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "integration: marks tests as integration tests (requiring external services)",
]
addopts = "-v --cov=. --cov-report=term-missing"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
